apps.core.models.assignment_group: Set fake deadlines to (1970,1,1). (#111)
It makes no sense to set it to datetime.now() because that only leads to
confusion. Good user interfaces should not show this deadline by date in
any case, so at least we should use a deadline that has virtually no
posibility of beeing misunderstood as a real one (by developers).
Furthermore, we can not set this field to None (null) since that will
break sorting.