apps.core.models.relateduser: Simplified #144.
To get a version of related users that we can improve through iteration,
I simplified RelatedUser by removing tag, and changing user from
ForeignKey to a username SlugField. For a more complex implementation
using user as a foreign key, we need to be able to set foreign keys
by other values than the primary key in simplified.