apps.examiner.views: Correct date and prefix usernames on groups for zip.
This partly fixes the requested features in issue #185, and closes #185.

The issue with naming is not possible to do in an inefficient
manner because users can be in more than one group on the same
assignment. Since users can be in more than one group on the same
assignment, they can also be the only student in more than one group
(and there are use cases where this is needed).

The deadline date was a bug, and it has been changed to
YYYY-MM-DDThh_mm_ss. Uses _ instead of : because it is more natural to
use from a command line interface.