devilry_theme: Update to bootstrap2.
Includes a much cleaner integration of bootstrap. The only thing shared
with the devilry ExtJS theme is the bootstrap variables. Other than
that, they are compiled into separate css files.

The only required changes seems to be something with list indent, and
using .muted instead of SMALL (which we should have done in any case).