devilry_settings: Override the default datetime formats.
Makes them less ambigous. This is important to avoid confusion with
non-native english speakers because 12/01/13 may mean "Jan 12" or "Dec 1"
depending on the ordering, and the english ordering is not universal. So
instead we use:

	Jan. 1, 2013 (short format)
	January 1, 2013 (long format)

We also use a 24 hour clock instead of AM and PM.

I think this will make dates and times easier to read and harder to
misunderstand.